refactor: move to subComponents folder #37
