+++ src/Components/Common/Card/subcomponents/CardBack/CardBack.style.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { CardContainer, rotate, rotateInverse } from '../Card.style';
+import { CardContainer, rotate, rotateInverse } from '../../Card.style';
 
 export const CardBackContainer = styled(CardContainer)<{ isCorrect: boolean }>`
   background-color: ${(props) =>
