+++ src/Components/Common/Card/subcomponents/CardFront/CardFront.style.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { CardContainer, rotate, rotateInverse } from '../Card.style';
+import { CardContainer, rotate, rotateInverse } from '../../Card.style';
 
 export const CardFrontContainer = styled(CardContainer)`
   animation-name: ${(props) => {
