+++ src/Components/Common/Card/subcomponents/CardFront/index.ts
