+++ src/Components/Common/Card/subcomponents/index.ts
@@ -1,0 +1,3 @@
+export * from './EditCard';
+export * from './CardBack';
+export * from './CardFront';
